+++ connector-application/src/test/java/eu/ecodex/connector/application/service/auth/token/ConnectorRegisterUserRefreshTokenServiceTest.java
@@ -23,7 +23,7 @@ import eu.ecodex.connector.application.exception.ConnectorUserBadCredentialsExce
 import eu.ecodex.connector.application.port.spi.auth.token.ConnectorAuthenticationTokenProvider;
 import eu.ecodex.connector.application.port.spi.auth.token.ConnectorRefreshTokenRepository;
 import eu.ecodex.connector.domain.model.auth.ConnectorRefreshToken;
-import eu.ecodex.connector.domain.model.login.ConnectorLoginResponse;
+import eu.ecodex.connector.application.port.spi.auth.login.ConnectorLoginResponse;
 import eu.ecodex.connector.domain.model.user.ConnectorUser;
 import java.time.Clock;
 import java.time.Duration;