+++ connector-integrationtest/src/test/java/eu/ecodex/connector/rest/admin/role/ConnectorUpdateRoleIT.java
@@ -117,5 +117,4 @@ class ConnectorUpdateRoleIT extends AbstractIntegrationTest {
             .expectStatus()
             .is4xxClientError();
     }
-
 }