+++ connector-integrationtest/src/test/java/eu/ecodex/connector/rest/auth/login/ConnectorRefreshTokenCleanupSchedulerIT.java
@@ -40,11 +40,9 @@ public class ConnectorRefreshTokenCleanupSchedulerIT extends AbstractIntegration
         cleanDb();
     }
 
-
     @Test
     @Sql("classpath:sql/user.sql")
     void should_clean_revoked_refresh_tokens() {
-
         var userUuid = "d43bfa931-3c25-47e4-b377-bf4ce7b0d04c_default_admin";
         var revoked = refreshTokenRepository.findByUserUuidAndRevoked(userUuid, true);
         var notRevoked = refreshTokenRepository.findByUserUuidAndRevoked(userUuid, false);